// my-app/src/enums/Color.ts
export enum Color {
    Red = 'red',
    Green = 'green',
    Blue = 'blue',
    Yellow = 'yellow',
    Purple = 'purple',
    Default = 'default',
    Primary = 'primary',
    Secondary = 'secondary',
    Error = 'error',
    Info = 'info',
    Success = 'success',
    Warning = 'warning'
} 

export enum ColorWithOpacity {
    Red = 'rgba(220, 20, 60, 0.5)',
    Green = 'rgba(50, 205, 50, 0.5)',
    Blue = 'rgba(100, 149, 237, 0.5)',
    Yellow = 'rgba(255, 225, 28, 0.5)',
    Purple = 'rgba(122, 61, 227, 0.5)'
}